+++ Test/UnitTest/BuiltInTest_Cpp/BuiltInTest_Cpp.vcxproj
@@ -159,10 +159,12 @@
     <ClCompile Include="..\CommonMain.cpp" />
     <ClCompile Include="TestLexer.cpp" />
     <ClCompile Include="TestStatements.cpp" />
+    <ClCompile Include="TestTraced.cpp" />
     <ClCompile Include="TestTypeExpr.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\..\Source\LogTrace.h" />
+    <ClInclude Include="TestCppHelper.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
